mv WFCondTask impl into WFCondTaskFactory.cc and override send()
